pyproject.toml
[tool.poetry]
name = "braniaclms"
version = "0.1.0"
description = ""
authors = ["bellagrin10 <bellgrinberg@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
requests = "^2.27.1"
python-dotenv = "^0.20.0"
ipython = "^8.4.0"
django-seeding-data = "^0.1.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.10.1"
black = {version = "^22.3.0", allow-prereleases = true}

[tool.black]
line-length = 79

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
